+++ src/types.rs
@@ -8,6 +8,7 @@ mod component;
 mod convert;
 mod convert_util;
 mod from_component;
+mod from_register_request;
 mod from_renderer;
 mod register_request;
 mod renderer;
